改变淘宝id字段类型

+++ src/main/java/com/imzhitu/admin/trade/item/pojo/Item.java
@@ -89,7 +89,7 @@ public class Item extends AbstractNumberDto implements Serializable {
 	 * 淘宝商品真实id
 	 * @author zhangbo	2015年12月9日
 	 */
-	private Integer itemId;
+	private Long itemId;
 	
 	/**
 	 * 淘宝物品:1,天猫:2
@@ -287,17 +287,12 @@ public class Item extends AbstractNumberDto implements Serializable {
 		this.stock = stock;
 	}
 
-	/**
-	 * @return the itemId
-	 */
-	public Integer getItemId() {
+
+	public Long getItemId() {
 		return itemId;
 	}
 
-	/**
-	 * @param itemId the itemId to set
-	 */
-	public void setItemId(Integer itemId) {
+	public void setItemId(Long itemId) {
 		this.itemId = itemId;
 	}
